remove trait
