+++ packages/fuels-accounts/src/kms/google/signer.rs
@@ -1,7 +1,7 @@
-use crate::kms::kms_trait::KmsSigner;
 use crate::kms::signature_utils::{convert_to_fuel_signature, normalize_signature};
 use async_trait::async_trait;
 use fuel_crypto::{Message, PublicKey, Signature};
+use fuels_core::traits::Signer;
 use fuels_core::types::{
     bech32::{Bech32Address, FUEL_BECH32_HRP},
     errors::{Error, Result},
@@ -139,12 +139,8 @@ impl GoogleKmsSigner {
 }
 
 #[async_trait]
-impl KmsSigner for GoogleKmsSigner {
-    fn fuel_address(&self) -> &Bech32Address {
-        &self.fuel_address
-    }
-
-    async fn sign_message(&self, message: Message) -> Result<Signature> {
+impl Signer for GoogleKmsSigner {
+    async fn sign(&self, message: Message) -> Result<Signature> {
         let signature_der = self.request_gcp_signature(message).await?;
 
         let k256_key = K256PublicKey::from_public_key_pem(&self.public_key_pem).map_err(|_| {
@@ -157,6 +153,10 @@ impl KmsSigner for GoogleKmsSigner {
             normalize_signature(&signature_der, message, &k256_key, GOOGLE_KMS_ERROR_PREFIX)?;
 
         Ok(convert_to_fuel_signature(normalized_sig, recovery_id))
+    }
+
+    fn address(&self) -> &Bech32Address {
+        &self.fuel_address
     }
 }
